Updated formatting / styles in all files
Added >Schedule to allow moderators an interactive way to schedule playtests. This will automatically remove the request from the spreadsheet, and add it to the calendar. It also allows for changing some required fields and data validation.

>P post will now actually disable voice

Adde >CallAllTesters to mention the regular playtester role when trying to get more players for a comp test.

>pa now includes help text for telling users how to get community tester

Added playtest request data model

Update OAuth to better store credentials in a central location. Along with ensuring credentials on startup.

Updated indexs in playtest event to no longer expect a featured images.

Added CheckForScheduleConflict() to allow checking if a test event is scheduled on a specific day.

sv_cheats should now be reset to 0 before a test.

Added sheets service for interaction with the playtesting queue.

Updated rcon reply string formatting.

Registered servers in Program.cs